requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
PyMySQL==1.1.1
bcrypt==4.2.0
python-dotenv==1.0.1
APScheduler==3.10.4
Werkzeug==3.0.3
